All SQLITE_MAX_VARIABLE_NUMBER to exceed 32767.  The sizes of some structures increase when the compile-time parameter is configured this way.
FossilOrigin-Name: 16a24b4485420bdf30d3c8e22cfbaf008e07df02

+++ src/vdbeaux.c
@@ -1359,7 +1359,7 @@ void sqlite3VdbeMakeReady(
 
     p->nCursor = (u16)nCursor;
     if( p->aVar ){
-      p->nVar = (u16)nVar;
+      p->nVar = nVar;
       for(n=0; n<nVar; n++){
         p->aVar[n].flags = MEM_Null;
         p->aVar[n].db = db;
